pyproject.toml
[project]
name = "real-estate-be"
version = "0.1.0"
description = "Litestar Backend for Real Estate Application"
authors = [
    {name = "KawaiButa",email = "21522441@gm.uit.edu.vn"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "litestar[standard, jwt, prometheus, structlog] (>=2.15.1,<3.0.0)",
    "pydantic (==2.10.6)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "geoalchemy2 (>=0.17.1,<0.18.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "advanced-alchemy (>=0.34.0,<0.35.0)",
    "supabase (>=2.13.0,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pygeohash (>=3.0.1,<4.0.0)",
    "faker (>=37.1.0,<38.0.0)",
    "openai (>=1.69.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "qrcode (>=8.1,<9.0)",
    "pillow (>=11.1.0,<12.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
